feat: 完全分離アーキテクチャ実装
