+++ kadai7_database-integraiton/openapi/schemas/event.yaml
@@ -1,0 +1,27 @@
+Event:
+  type: object
+  properties: 
+    id:
+      type: integer
+      example: 1
+      description: イベントID
+    date:
+      type: string
+      format: date
+      example: "2025-09-25"
+      description: イベント日時
+    title:
+      type: string
+      example: "@JAM EXPO 2025 supported by UP-T"
+      description: イベントタイトル
+    isAttending:
+      type: boolean
+      example: false
+      description: 参加予定フラグ
+    details:
+      type: object
+      additionalProperties:
+        type: string
+      example:
+        description: ""
+      description: イベント詳細情報（現在は空の場合が多い）
